refactor : 알파벳 카드 클래스 분리
