+++ README.md
@@ -6,9 +6,9 @@
 * [ ] 게임 결과를 출력한다.
 * [ ] 게임 시작시 참여자는 카드를 2장 씩 받는다.
 * [x] 숫자 카드는 숫자를 점수로 한다.
-* [ ] ace 는 1 또는 11을 점수로 계산한다.
+* [x] ace 는 1 또는 11을 점수로 계산한다.
 * [ ] king, queen, jack 은 10을 점수로 계산한다.
-* [x] 카드 숫자는 2~10,king,queen,jack,ace 총 13개이다.
+* [ ] 카드 숫자는 2~10,king,queen,jack,ace 총 13개이다.
 * [x] 카드 문양은 총 4가지이다. ( 클로버, 하트, 다이아, 스페이드 )
 * [ ] 카드의 총 개수는 52장이다.
 * [ ] 점수의 합이 21을 넘지 않는 동안 카드를 계속 받을 수 있다.
