+++ src/main/kotlin/blackjack/NumberCard.kt
@@ -1,0 +1,9 @@
+package blackjack
+
+data class NumberCard(override val symbol: Symbol, val number: Int) : Card {
+    init {
+        require(number in 2..10) { "숫자 카드의 범위는 2 ~ 10 입니다." }
+    }
+
+    override fun score(): Int = number
+}
